+++ src/jpeg2000/coding_parameters.cc
@@ -1,8 +1,15 @@
 #include "coding_parameters.h"
 
 #include <cstdint>
+#include <cstdlib>
 
 namespace jpeg2000 {
+
+    Size CodingParameters::SizeAtLevel(int level) const {
+        uint64_t scale = uint64_t(1) << level;
+        return Size(DivideRoundUp(size.x, scale),
+                    DivideRoundUp(size.y, scale));
+    }
 
     int CodingParameters::GetPositionIndex(int r, int px, int py, int *position_resolutions,
                                            int *resolution_at_position) const {
@@ -15,9 +22,9 @@ namespace jpeg2000 {
         for (int i = 0; i <= num_levels; ++i) {
             int64_t step_x = static_cast<int64_t>(resolutions[i].precinct_size.x) << (num_levels - i);
             int64_t step_y = static_cast<int64_t>(resolutions[i].precinct_size.y) << (num_levels - i);
-            Size precincts = GetPrecincts(i, size);
 
-            position_index += static_cast<int>((y + step_y - 1) / step_y) * precincts.x;
+            position_index += static_cast<int>((y + step_y - 1) / step_y) *
+                              resolutions[i].num_precincts.x;
             if (y % step_y != 0)
                 continue;
 
@@ -34,83 +41,52 @@ namespace jpeg2000 {
     void CodingParameters::FillPrecinctCounts() {
         total_precincts = 0;
         for (int i = 0; i <= num_levels; ++i) {
-            resolutions[i].first_precinct = total_precincts;
-            Size precincts = GetPrecincts(i, size);
-            total_precincts += precincts.x * precincts.y;
+            Resolution &resolution = resolutions[i];
+            resolution.first_precinct = total_precincts;
+            resolution.num_precincts = GetPrecincts(i, size);
+            total_precincts += resolution.num_precincts.x * resolution.num_precincts.y;
         }
     }
 
     int CodingParameters::GetClosestResolution(const Size &res_size, Size *res_image_size) const {
-        int distance, final_r = 0;
-        int distance_x = size.x - res_size.x;
-        int distance_y = size.y - res_size.y;
-        int min = abs(distance_x) + abs(distance_y);
-        int res_image_x, res_image_y;
-        res_image_size->x = size.x;
-        res_image_size->y = size.y;
+        int final_level = 0;
+        int64_t distance_x = static_cast<int64_t>(size.x) - res_size.x;
+        int64_t distance_y = static_cast<int64_t>(size.y) - res_size.y;
+        int64_t minimum = std::abs(distance_x) + std::abs(distance_y);
+        *res_image_size = size;
 
-        for (int r = 1; r <= num_levels; ++r) {
-            res_image_x = (int) ceil((double) size.x / (1L << r));
-            res_image_y = (int) ceil((double) size.y / (1L << r));
-            distance_x = res_image_x - res_size.x;
-            distance_y = res_image_y - res_size.y;
-            distance = abs(distance_x) + abs(distance_y);
+        for (int level = 1; level <= num_levels; ++level) {
+            Size level_size = SizeAtLevel(level);
+            distance_x = static_cast<int64_t>(level_size.x) - res_size.x;
+            distance_y = static_cast<int64_t>(level_size.y) - res_size.y;
+            int64_t distance = std::abs(distance_x) + std::abs(distance_y);
 
-            if (distance < min) {
-                res_image_size->x = res_image_x;
-                res_image_size->y = res_image_y;
-                min = distance;
-                final_r = r;
+            if (distance < minimum) {
+                *res_image_size = level_size;
+                minimum = distance;
+                final_level = level;
             }
         }
-
-        int res = (num_levels - final_r);
-
-        if (res > num_levels) res = num_levels;
-        else if (res < 0) res = 0;
-
-        return res;
+        return num_levels - final_level;
     }
 
     int CodingParameters::GetRoundUpResolution(const Size &res_size, Size *res_image_size) const {
-        int r = num_levels;
-        bool bigger = false;
-
-        while (!bigger && r >= 0) {
-            res_image_size->x = (int) ceil((double) size.x / (1L << r));
-            res_image_size->y = (int) ceil((double) size.y / (1L << r));
-            if ((res_image_size->x >= res_size.x) &&
-                (res_image_size->y >= res_size.y))
-                bigger = true;
-            else r--;
+        for (int level = num_levels; level >= 0; --level) {
+            *res_image_size = SizeAtLevel(level);
+            if (res_image_size->x >= res_size.x && res_image_size->y >= res_size.y)
+                return num_levels - level;
         }
-
-        int res = (num_levels - r);
-
-        if (res > num_levels) res = num_levels;
-        else if (res < 0) res = 0;
-
-        return res;
+        *res_image_size = size;
+        return num_levels;
     }
 
     int CodingParameters::GetRoundDownResolution(const Size &res_size, Size *res_image_size) const {
-        int r = 0;
-        bool smaller = false;
-
-        while (!smaller && r <= num_levels) {
-            res_image_size->x = (int) ceil((double) size.x / (1L << r));
-            res_image_size->y = (int) ceil((double) size.y / (1L << r));
-            if ((res_image_size->x <= res_size.x) &&
-                (res_image_size->y <= res_size.y))
-                smaller = true;
-            else r++;
+        for (int level = 0; level <= num_levels; ++level) {
+            *res_image_size = SizeAtLevel(level);
+            if (res_image_size->x <= res_size.x && res_image_size->y <= res_size.y)
+                return num_levels - level;
         }
-
-        int res = (num_levels - r);
-
-        if (res > num_levels) res = num_levels;
-        else if (res < 0) res = 0;
-
-        return res;
+        *res_image_size = SizeAtLevel(num_levels);
+        return 0;
     }
 }